+++ source/dotnet/Measurements.UnitTests/Infrastructure/Handlers/MeasurementsHandlerTests.cs
@@ -85,7 +85,7 @@ public class MeasurementsHandlerTests
         Assert.Equal(42, actualAggregations.Quantity);
         Assert.Equal(Quality.Measured, actualAggregations.Quality);
         Assert.Equal(Unit.kWh, actualAggregations.Unit);
-        Assert.True(actualAggregations.MissingValues);
+        Assert.True(actualAggregations.ContainsMissingValues);
         Assert.True(actualAggregations.ContainsUpdatedValues);
     }